refactor(data): 删除 mbse.data.enums 兼容 shim，清理指向已删旧包的死引用
- 枚举真身早在 e2m2e.data.templates.enums，迁移 tests/ 下 6 个文件的
  调用点后删除 shim 文件与专门测 shim 的 test_old_paths_shim
- mbse/data/__init__.py 不再重导出枚举；docs/api/e2m2e.mbse.rst 同步
- 修正声称旧路径（e2m2e.core.*）经 shim 可用的过时注释（第 5 批清理
  已完成，shim 早已不存在）：exceptions.py、data/__init__.py、
  templates/enums.py、kernels/manager.py
- transfer/ 下 11 个文件 docstring 的 Sphinx 引用 e2m2e.transfer.*
  改为 e2m2e.algorithm.transfer.*
