+++ docs/api/e2m2e.mbse.rst
@@ -22,17 +22,6 @@ e2m2e.mbse.data.core_models module
    :exclude-members: ProjectionPlane
 
 
-e2m2e.mbse.data.enums module
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: e2m2e.mbse.data.enums
-   :no-index:
-
-   ADR 0011 迁移后的领域枚举 shim：全部枚举从 ``e2m2e.data.templates.enums``
-   重导出，此处仅作兼容路径，不重复登记成员（避免与正式定义产生交叉引用歧义）。
-   枚举定义见 :mod:`e2m2e.data.templates.enums`。
-
-
 e2m2e.mbse.architecture package
 -------------------------------
