+++ e2m2e/mbse/data/__init__.py
@@ -1,34 +1,12 @@
 """数据模型子包
 
-提供基于 Pydantic 的统一数据结构。
+提供基于 Pydantic 的统一数据结构。领域枚举见 ``e2m2e.data.templates.enums``。
 """
 
 from .core_models import (
     OrbitProperties,
 )
-from .enums import (
-    BifurcationLabel,
-    BoundaryMode,
-    ConvergenceState,
-    OrbitFamilyType,
-    ProjectionPlane,
-    ReferenceFrame,
-    StabilityLabel,
-    TransferPhase,
-    TransferType,
-    UnitSystem,
-)
 
 __all__ = [
-    "BifurcationLabel",
-    "BoundaryMode",
-    "ConvergenceState",
-    "OrbitFamilyType",
-    "ProjectionPlane",
-    "ReferenceFrame",
-    "StabilityLabel",
-    "TransferPhase",
-    "TransferType",
-    "UnitSystem",
     "OrbitProperties",
 ]
