+++ tests/algorithm/transfer/test_dro_ro_nlp.py
@@ -14,8 +14,8 @@ from e2m2e.algorithm.transfer.transfer_optimization import (
     NLPOptimizationVariables,
 )
 from e2m2e.data.templates import ConvergenceState, FailureCause
+from e2m2e.data.templates.enums import TransferType
 from e2m2e.data.types.orbit import Orbit
-from e2m2e.mbse.data.enums import TransferType
 
 pytestmark = pytest.mark.orchestration
